Removed count on CustomList
